examples/sample-app-python/requirements.txt
Flask==3.1.0
prometheus-client==0.21.0
opentelemetry-api==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-exporter-otlp-proto-grpc==1.29.0
opentelemetry-instrumentation-flask==0.50b0
opentelemetry-semantic-conventions==0.50b0
